+++ packages/styles/src/packages/index.ts
@@ -1,2 +1,3 @@
-export * from "./title.css";
+export * from "./skeleton.css";
 export * from "./text.css";
+export * from "./title.css";